+++ src/common/init.c
@@ -163,9 +163,6 @@ void ffListFeatures(void)
         #ifdef FF_HAVE_THREADS
             "threads\n"
         #endif
-        #ifdef FF_HAVE_LIBPCI
-            "libpci\n"
-        #endif
         #ifdef FF_HAVE_VULKAN
             "vulkan\n"
         #endif
@@ -240,6 +237,9 @@ void ffListFeatures(void)
         #endif
         #ifdef FF_HAVE_DDCUTIL
             "libddcutil\n"
+        #endif
+        #if __has_include(<linux/videodev2.h>)
+            "videodev2\n"
         #endif
         #ifdef FF_HAVE_DIRECTX_HEADERS
             "Directx Headers\n"