+++ src/detection/camera/camera.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "fastfetch.h"
+
+typedef struct FFCameraResult
+{
+    FFstrbuf name;
+    FFstrbuf vendor;
+    FFstrbuf id;
+    FFstrbuf colorspace;
+    uint32_t width;
+    uint32_t height;
+} FFCameraResult;
+
+const char* ffDetectCamera(FFlist* result /* list of FFCameraResult */);