+++ src/detection/terminalshell/terminalshell_linux.c
@@ -285,6 +285,9 @@ static pid_t getTerminalInfo(FFTerminalResult* result, pid_t pid)
             ffStrEquals(name, "xonsh")      || // works in Linux but not in macOS because kernel returns `Python` in this case
             ffStrEquals(name, "login")      ||
             ffStrEquals(name, "sshd")       ||
+            #ifdef __linux__
+            ffStrStartsWith(name, "flatpak-") || // #707
+            #endif
             ffStrEndsWith(name, ".sh")
         )
         {