+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 2.8.0
+
+Changes:
+* The linux binaries are now built with glibc 2.35, which means they no longer support Debian 11 and Ubuntu 20.04. Users using these distros may download the artifacts `fastfetch-linux-old` from GitHub Actions.
+
+Features:
+* Rewrite GPU module, drop libpci dependency (GPU, Linux)
+* Detect marketing name of Apple Silicon CPUs for asahi linux (CPU, Linux)
+* Add new module `Camera`, which prints the name and resolution of connected cameras
+
+Bugfixes:
+* Fix compatibility with packages installed by flatpak (Terminal, Linux)
+* Don't show an empty battery if no battery is detected (macOS, Battery)
+* Don't show `not connected` if no power adapter is found (macOS / Linux, PowerAdapter)
+* Make format of battery status be consistent with other platforms (Linux, Battery)
+
+Logo:
+* Print Asahi logo in asahi linux (Logo, Linux)
+* Add Asahi2, z/OS, Tatra, PikaOS
+
 # 2.7.1
 
 Features: